[android][sqlite] Use 64-bit IEEE 754 floats to match SQLite
`REAL` values in SQLite are 64-bit floats: https://www.sqlite.org/datatype3.html. Need to use the same type of data type here -- 64-bit, IEEE 754 floating point.

Should fix #3013

+++ android/expoview/src/main/java/versioned/host/exp/exponent/modules/api/SQLiteModule.java
@@ -160,7 +160,7 @@ public class SQLiteModule extends ReactContextBaseJavaModule {
   private Object getValueFromCursor(Cursor cursor, int index, int columnType) {
     switch (columnType) {
       case Cursor.FIELD_TYPE_FLOAT:
-        return cursor.getFloat(index);
+        return cursor.getDouble(index);
       case Cursor.FIELD_TYPE_INTEGER:
         return cursor.getInt(index);
       case Cursor.FIELD_TYPE_BLOB:
